% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatebydate.R
\name{AggregateByDate}
\alias{AggregateByDate}
\title{Aggregate values by date}
\usage{
AggregateByDate(x, by, dates, FUN = sum, ...)
}
\arguments{
\item{x}{A vector containing the values to be aggregated, or, a 1D-array,
or some other structure with dates in the first column}

\item{by}{The period used in the conversion (e.g., "week", "year").}

\item{dates}{Either dates or characters that can be coerced to dates.
Inferred from x if not suppled (e.g., the names of x).}

\item{FUN}{The function to be applied, or NULL. In the case of functions
like +, %*%, etc., the function name must be backquoted or quoted.
If FUN is NULL, tapply returns a vector which can be used to subscript the
multi-way array tapply normally produces.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A vector of aggregated values.
}
\description{
Aggregates values by date. Calls \code{\link{tapply}}.
}
\examples{
z <- rep(1, 54)
dts <- seq.Date(as.Date("2017/01/01"), by = "week", length.out = length(z))
AggregateByDate(z, dates = dts, by = "year")

z <- seq_along(dts)
names(z) <- dts
AggregateByDate(z, by = "year", FUN = mean)

data(presidents, package = "datasets")
origin <- paste0(start(presidents)[1], "-01-01")
df <- data.frame(dates = seq.Date(as.Date(0, origin), length.out = length(presidents),
                                                           by = "quarters"),
                             x = as.numeric(presidents))
AggregateByDate(df, by = "year", FUN = median, na.rm = TRUE)
}
