% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{Change29FebTo28th}
\alias{Change29FebTo28th}
\title{Change29FebTo28th}
\usage{
Change29FebTo28th(x)
}
\arguments{
\item{x}{Vector of dates}
}
\value{
A vector of dates
}
\description{
Change29FebTo28th
}
\details{
Replaces any 29th of Februaries with 28th of Februaries.
}
