% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{CompleteListPeriodNames}
\alias{CompleteListPeriodNames}
\title{\code{CompleteListPeriodNames}}
\usage{
CompleteListPeriodNames(x, by)
}
\arguments{
\item{x}{The vector of \code{char} to convert.}

\item{by}{The time aggregation.}
}
\description{
Returns a vector that contains all the possible period dates.
}
