% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{DaysPerPeriod}
\alias{DaysPerPeriod}
\title{\code{DaysPerPeriod}}
\usage{
DaysPerPeriod(by)
}
\arguments{
\item{by}{The period used in the conversion (e.g., "week", "year").}
}
\description{
The average number of dates in a period. E.g., 7 for a day, 365.25 for a year.
}

