% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{DecimalDate}
\alias{DecimalDate}
\title{DecimalDate}
\usage{
DecimalDate(x, by)
}
\arguments{
\item{x}{Date}

\item{by}{The period used in the conversion. Either "month" or "year".}
}
\description{
DecimalDate
}
\details{
Turns a date into a integer, relative to an origin of 1900-01-01. Ignores hours, minutes and seconds.
}
\examples{
DecimalDate("2000/01/01", by = "year")
DecimalDate("2000/01/01", by = "month")
DecimalDate("2000/02/01", by = "month")
DecimalDate("2000/06/16", by = "month")
}
