% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{DiffPeriod}
\alias{DiffPeriod}
\title{\code{DiffYear}}
\usage{
DiffPeriod(from, to, by, ceiling = FALSE)
}
\arguments{
\item{from}{The earlier date.}

\item{to}{The later date.}

\item{by}{The period used in the conversion. Either "month" or "year".}

\item{ceiling}{If TRUE, rounds partial-year differences up (otherwise they are rounded down).}
}
\value{
An integer.
}
\description{
Computes the difference in years between two dates.
}
\examples{
DiffPeriod("2015/05/06", "2017/05/06", by = "year")
}
\references{
Adapted from Brian Ripley, http://r.789695.n4.nabble.com/Calculate-difference-between-dates-in-years-td835196.html
}
