% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{FormatPeriod}
\alias{FormatPeriod}
\title{FormatPeriod}
\usage{
FormatPeriod(period)
}
\arguments{
\item{period}{A period object. Alternatively a numeric corresponding to a
number of seconds.}
}
\value{
A string displaying the period:
e.g. "1 Day, 2 Hours, 3 Minutes and 4 Seconds"
}
\description{
Displays a period in days, hours, minutes and seconds.
}
\examples{
FormatPeriod(1000)
}
