% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{ParseDates}
\alias{ParseDates}
\title{{ParseDates}}
\usage{
ParseDates(x, us.format = NULL)
}
\arguments{
\item{x}{Character vector to be parsed.}

\item{us.format}{Whether to use the US convention for dates.}
}
\description{
Parses date characters.
}
\examples{
ParseDates("1-2-2017", us.format = FALSE)
}

