% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{Period}
\alias{Period}
\title{\code{Period}}
\usage{
Period(x, by)
}
\arguments{
\item{x}{The date.}

\item{by}{The period used in the conversion (e.g., "week", "year").}
}
\description{
Converts a date into a character.
}

