% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{PeriodNameToDate}
\alias{PeriodNameToDate}
\title{\code{PeriodNameToDate}}
\usage{
PeriodNameToDate(x, by)
}
\arguments{
\item{x}{The vector of \code{char} to convert.}

\item{by}{The time aggregation.}
}
\description{
Converts a vector of period names in format
yyyy, yyyy-mm-dd, or yyyy-mm into a Date.
}
\examples{
PeriodNameToDate(2010:2014, "year")
PeriodNameToDate(c("2010-01", "2010-02"), "month")
}

