% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{PeriodNameToYMD}
\alias{PeriodNameToYMD}
\title{\code{PeriodNameToYMD}}
\usage{
PeriodNameToYMD(x)
}
\arguments{
\item{x}{The vector of \code{char} to convert.}
}
\description{
Converts a vector of period names in format
yyyy, yyyy-mm-dd, yyyy-mm,  into a Date.
}
\examples{
PeriodNameToYMD(2010:2014)
PeriodNameToYMD(c("2010-01", "2010-02"))
PeriodNameToYMD(c("26/02/2011-1/01/2012", "2/01/2012-8/01/2012"))
}

