% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{Periods}
\alias{Periods}
\title{\code{Periods}}
\usage{
Periods(x = 1, by)
}
\arguments{
\item{x}{A numeric value of the number of units to be contained in the period. With the exception of seconds(), x must be an integer.}

\item{by}{The period used in the conversion (e.g., "week", "year").}
}
\description{
Quickly cretes period objects.
}
