% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{RefreshIfOlderThan}
\alias{RefreshIfOlderThan}
\title{{RefreshIfOlderThan}}
\usage{
RefreshIfOlderThan(x, units = "seconds")
}
\arguments{
\item{x}{The period expressed in \code{units} units.}

\item{units}{The time unit. One of \code{"seconds"}, \code{"minutes"}, \code{"hours"}, \code{"days"},
\code{"weeks"} or \code{"months"}.}
}
\description{
An R object will be automatically recalculated if its output is older than the duration specified.
Recalculation will only occur when someone looks at that R object, or something that is calculated using it.
}
\details{
If \code{units} = \code{"months"}, then \code{x} must be an integer.
}
\examples{
RefreshIfOlderThan(2, "hours")
RefreshIfOlderThan(1, "weeks")
}
