% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriescolumnchart.R
\name{TimeSeriesColumnChart}
\alias{TimeSeriesColumnChart}
\title{\code{TimeSeriesColumnChart}}
\usage{
TimeSeriesColumnChart(x, title = "", xtitle = "", ytitle = "",
  series.name = "", tickformat = NULL, smooth = TRUE,
  ignore.last.period.from.smooth = FALSE)
}
\arguments{
\item{x}{A vector containing values, where the names indicate the dates.}

\item{title}{The title of the chart.}

\item{xtitle}{The title to show on the x-axis.}

\item{ytitle}{The title to show on the y-axis.}

\item{series.name}{E.g., "churn".}

\item{tickformat}{Plotlytickformat \code{tickformat}.}

\item{smooth}{Smooth the data using \code{supmsu}.}

\item{ignore.last.period.from.smooth}{If \code{TRUE}, the last period is ignored from the smoothing.}
}
\value{
A plotly plot.
}
\description{
Plots a time series as columns, optionally smoothing the data.
}
\examples{
z = runif(9)
names(z) = seq(as.Date("2010/1/1"), as.Date("2018/1/1"), "years")
TimeSeriesStackedColumnChart(z, "year")
}
