% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesstackedcolumn.R
\name{TimeSeriesStackedColumnChart}
\alias{TimeSeriesStackedColumnChart}
\title{\code{TimeSeriesStackedColumnChart}}
\usage{
TimeSeriesStackedColumnChart(x, by = "day", title = "", ytitle = "",
  xtitle = "", series.name = "", tickformat = NULL)
}
\arguments{
\item{x}{table with dates in the columns.}

\item{by}{The period that has been used to aggregate the data (day, week, month, quarter, year).}

\item{title}{The title of the chart.}

\item{ytitle}{The title to show on the y-axis.}

\item{xtitle}{The title to show on the x-axis.}

\item{series.name}{E.g., "churn".}

\item{tickformat}{Plotlytickformat \code{tickformat}.}
}
\value{
A plotly plot.
}
\description{
Plots a time series as columns.
}
\examples{
z = matrix(runif(27), nrow = 3, dimnames = list(LETTERS[1:3],
seq(as.Date("2010/1/1"), as.Date("2018/1/1"), "years")))
TimeSeriesStackedColumnChart(z, "year")
}
