% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upadting.R
\name{TimeUnitsToSeconds}
\alias{TimeUnitsToSeconds}
\title{{TimeUnitsToSeconds}}
\usage{
TimeUnitsToSeconds(x, units = "seconds")
}
\arguments{
\item{x}{The period expressed in \code{units} units.}

\item{units}{The time unit, which can be seconds, minutes, days, weeks or months.}
}
\description{
Converts a number of minutes, hours, days, weeks or months to seconds.
}
