% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upadting.R
\name{UpdateIn}
\alias{UpdateIn}
\title{{UpdateIn}}
\usage{
UpdateIn(x, units = "seconds")
}
\arguments{
\item{x}{The period expressed in \code{units} units.}

\item{units}{The time unit, which can be seconds, minutes, days, weeks or months.}
}
\description{
Sets a period of time, after which an R object is woken and updated.
}
\examples{
UpdateIn(5, "days")
UpdateIn(1, "months")
}
