% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatebydate.R
\name{AggregateByDate}
\alias{AggregateByDate}
\title{\code{AggregateByDate}}
\usage{
AggregateByDate(x, by, dates = names(x), FUN = sum)
}
\arguments{
\item{x}{A vector containing the values to be aggregated.}

\item{by}{The period used in the conversion (e.g., "week", "year").}

\item{dates}{Either dates or characters that can be coerced to dates.}

\item{FUN}{The function to be applied, or NULL. In the case of functions
like +, %*%, etc., the function name must be backquoted or quoted.
If FUN is NULL, tapply returns a vector which can be used to subscript the
multi-way array tapply normally produces.}
}
\value{
A vector of aggregated values.
}
\description{
Aggregates values by date. Calls \code{\link{tapply}}.
}
\examples{
z <- rep(1, 54)
dts <- seq.Date(as.Date("2017/01/01"), by = "week", length.out = length(z))
AggregateByDate(z, dates = dts, by = "year")
}
