% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{ParseDates}
\alias{ParseDates}
\alias{AsDate}
\title{Parse Character Dates To POSIXct Objects}
\usage{
ParseDates(x, us.format = NULL)

AsDate(x, us.format = TRUE)
}
\arguments{
\item{x}{character; vector to be parsed}

\item{us.format}{logical; whether to use the US convention for dates; can be \code{NULL}
in which case both U.S. formats and international formats will be checked}
}
\value{
a vector of POSIXct date-time objects
}
\description{
Parse dates assuming some common (unknown) formats
popular either in the U.S. or internationally
}
\details{
\code{ParseDates} is deprecated and \code{AsDate}
should be preferred for efficiency reasons

\code{AsDate} does not allow \code{NULL} value for \code{us.format}
}
\examples{
ParseDates("1-2-2017", us.format = FALSE)
AsDate("1-2-2017", us.format = FALSE)
}
