% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{ParseDateTime}
\alias{ParseDateTime}
\alias{AsDateTime}
\title{Parse Character Date-Times to POSIXct Objects}
\usage{
ParseDateTime(x, us.format = TRUE, time.zone = "UTC")

AsDateTime(x, us.format = NULL, time.zone = "UTC", exact = TRUE)
}
\arguments{
\item{x}{Character vector to be parsed.}

\item{us.format}{Whether to use the US convention for dates.}

\item{time.zone}{An optional time zone, or else default of 'UTC' applies.}

\item{exact}{logical; see \code{\link[lubridate]{parse_date_time2}}; setting to \code{TRUE}
(the default) should result in slightly faster parsing, but there may be some cases that fail to parse correctly}
}
\value{
a vector of POSIXct date-time objects
}
\description{
Parses date-time character vectors to POSIXct
}
\details{
\code{ParseDateTime} is deprecated and merely calls \code{AsDateTime}
}
\examples{
AsDateTime("1-2-2017 12:34:56", us.format = FALSE)
}
\references{
See \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for a list of time zones.
}
\seealso{
\code{\link[lubridate]{parse_date_time2}}
}
