% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{DiffPeriod}
\alias{DiffPeriod}
\title{\code{DiffYear}}
\usage{
DiffPeriod(from, to, by, ceiling = FALSE)
}
\arguments{
\item{from}{The earlier date.}

\item{to}{The later date.}

\item{by}{The period used in the conversion. Either "month" or "year".}

\item{ceiling}{If TRUE, rounds partial-year differences up (otherwise they are rounded down).}
}
\value{
An integer.
}
\description{
Computes the difference in years between two dates.
}
\details{
Dates of "2016-02-29" are automatically converted to "2016-02-28" to deal with bugs in lubridate.
}
\examples{
DiffPeriod("2015/05/06", "2017/05/06", by = "year")
}
