% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdatetime.R
\name{IsDateTime}
\alias{IsDateTime}
\title{Check whether text can be parsed as date/time}
\usage{
IsDateTime(x)
}
\arguments{
\item{x}{Vector of input text}
}
\description{
Returns bool indicating whether text can be parsed as a date-time
}
\examples{
IsDateTime("2007")
IsDateTime("abc")
}
