% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{ParseDateTime}
\alias{ParseDateTime}
\title{\code{ParseDateTime}}
\usage{
ParseDateTime(x, us.format = TRUE)
}
\arguments{
\item{x}{Character vector to be parsed.}

\item{us.format}{Whether to use the US convention for dates.}
}
\description{
\code{ParseDateTime}
}
\examples{
ParseDateTime("1-2-2017 12:34:56", us.format = FALSE)
}

