% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{ParseDateTime}
\alias{ParseDateTime}
\title{{ParseDateTime}}
\usage{
ParseDateTime(x, us.format = TRUE, time.zone = "UTC")
}
\arguments{
\item{x}{Character vector to be parsed.}

\item{us.format}{Whether to use the US convention for dates.}

\item{time.zone}{An optional time zone, or else default of 'UTC' applies.
See https://en.wikipedia.org/wiki/List_of_tz_database_time_zones for a list of time zones.}
}
\description{
Parses date time characters.
}
\examples{
ParseDateTime("1-2-2017 12:34:56", us.format = FALSE)
}
