% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{Period}
\alias{Period}
\title{\code{Period}}
\usage{
Period(x, by, anchor.date = "1970-01-01")
}
\arguments{
\item{x}{The date.}

\item{by}{The period used in the conversion (e.g., "week", "year"). Special
cases include:
- "quarter", which provides the month for the quarter as YYYY-MM
- "nice.quarter", which provides the quarter in the form Q1 2022
- "2-week", "4-week", etc which provides multi-weekly periods labeled
  first date in the corresponding period. Requires to use to specify an
  anchor date as a point of reference.}

\item{anchor.date}{The date to use as a reference for multi-week periods. Can be a
Date object, or a string in the form "YYYY-MM-DD". For example,
when wanting two-week periods for "2022-07-04", does this date fall into
the 2 weeks commencing "2022-07-04" or the two weeks comencing "2022-06-27".}
}
\description{
Converts a date into a character.
}
