% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{PeriodNameToDate}
\alias{PeriodNameToDate}
\title{{PeriodNameToDate}}
\usage{
PeriodNameToDate(x, by, us.format = NULL)
}
\arguments{
\item{x}{The vector of \code{char} to convert.}

\item{by}{The time aggregation. Deprecated as this can be deduced from x.}

\item{us.format}{Whether to assume US date format when parsing.}
}
\description{
Converts a vector of period names in format
yyyy, yyyy-mm, yyyy-mm-dd, mmm-mmm yy, mmm yyyy, dd/mm/yyyy and
dd/mm/yyyy-dd/mm/yyyy into a Date.
}
\examples{
PeriodNameToDate(2010:2014)
PeriodNameToDate(c("2010-01", "2010-02"))
PeriodNameToDate(c("26/02/2011-1/01/2012", "2/01/2012-8/01/2012"))
}
