% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{UpdateEvery}
\alias{UpdateEvery}
\title{{UpdateEvery}}
\usage{
UpdateEvery(x, units = "seconds", wakeup = TRUE)
}
\arguments{
\item{x}{The period expressed in \code{units} units.}

\item{units}{The time unit, which can be seconds, minutes, days, weeks or months.}

\item{wakeup}{Whether to update even if document containing the R object is closed.}
}
\description{
Sets a period of time, after which an R object is woken and updated.
}
\details{
If \code{units} = "months" then \code{x} must be an integer. The update time
will roll back to the last day of the previous month if no such day exists \code{x} months
forward from today.
}
\examples{
UpdateEvery(5, "days")
UpdateEvery(1, "months")
}
